% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_mycobank_name.R
\name{get_current_mycobank_name}
\alias{get_current_mycobank_name}
\title{Get the 'current' Mycobank names for a vector of taxa names}
\usage{
get_current_mycobank_name(taxa, mycobank_db)
}
\arguments{
\item{taxa}{Character. Vector of taxa names to look up.}

\item{mycobank_db}{Data frame. Object name for Mycobank database. Typically created by get_mycobank_db()}
}
\value{
Data frame. Two columns: 1) query, the query of taxon names provided; 2) current_name, the current name given in the Mycobank database. Duplicated query names removed
}
\description{
Parses through the Mycobank synonyms and taxon names to pull the 'current' name for a list of taxa names
Names should be in a character vector and in the format "genus species" or just "genus" (variants and strains accepted)
If a given taxonomic name query consists of only one 'word' then taxonomic level of 'genus' is assumed
}
\details{
This will search through the database using pattern matching to pull the 'current name' so spelling counts. Capitalization does not affect it though.
}
\examples{
db <- get_mycobank_db(overwrite = FALSE)
taxa <- c("Abaphospora borealis","Conisphaeria borealis var. minor","Sphaeria borealis","Nonsense name","Abaphospora")
get_current_mycobank_name(taxa,db)

}
