% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mycobank_current_name.R
\name{get_mycobank_current_name}
\alias{get_mycobank_current_name}
\title{Get the 'current' accepted name from Mycobank for a vector of taxa names}
\usage{
get_mycobank_current_name(taxa, mycobank_db)
}
\arguments{
\item{taxa}{Character. Vector of taxa names to look up.}

\item{mycobank_db}{Data frame. Object name for Mycobank database. Typically created by get_mycobank_db()}
}
\value{
Character vector. This returns a vector of the "current names" for your query taxa.
}
\description{
Parses through the Mycobank synonyms and taxon names to pull the current accepted taxon name for a vector of query taxon names
Names should be in a character vector and in the format "genus species" or just "genus" (variants and strains accepted)
If a given taxonomic name query consists of only one 'word' then taxonomic level of 'genus' is assumed
}
\details{
This is a shortcut function that makes use of get_mycobank_synonyms() to just pull the "current name".
}
\examples{
db <- get_mycobank_db(overwrite = FALSE)
taxa <- c("Abaphospora borealis","Conisphaeria borealis var. minor","Sphaeria borealis","Nonsense name","Abaphospora")
get_mycobank_current_name(taxa,db)

}
