% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mycobank_synonyms.R
\name{get_mycobank_synonyms}
\alias{get_mycobank_synonyms}
\title{Get the 'current' Mycobank synonyms for a vector of taxa names}
\usage{
get_mycobank_synonyms(taxa, mycobank_db)
}
\arguments{
\item{taxa}{Character. Vector of taxa names to look up.}

\item{mycobank_db}{Data frame. Object name for Mycobank database. Typically created by get_mycobank_db()}
}
\value{
List. Contains 5 elements:  "query","current_name","basio_name","obligate_synonym","taxonomic_synonyms"
}
\description{
Parses through the Mycobank synonyms and taxon names to pull various synonyms for a vector of taxon names
Names should be in a character vector and in the format "genus species" or just "genus" (variants and strains accepted)
If a given taxonomic name query consists of only one 'word' then taxonomic level of 'genus' is assumed
}
\details{
This will search through the database using pattern matching to pull the various basionyms, and synonyms for a given taxon name, if present. Returns NA for that taxon otherwise.
}
\examples{
db <- get_mycobank_db(overwrite = FALSE)
taxa <- c("Abaphospora borealis","Conisphaeria borealis var. minor","Sphaeria borealis","Nonsense name","Abaphospora")
synonyms <- get_mycobank_synonyms(taxa,db)
# To see a tidy format for the first taxon query:
synonyms \%>\% map(1) \%>\% as.data.frame()

}
