% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mycobank_taxonomy.R
\name{get_mycobank_taxonomy}
\alias{get_mycobank_taxonomy}
\title{Get the Mycobank taxonomic classifications for a vector of taxa names}
\usage{
get_mycobank_taxonomy(taxa, mycobank_db)
}
\arguments{
\item{taxa}{Character. Vector of taxa names to look up.}

\item{mycobank_db}{Data frame. Object name for Mycobank database. Typically created by get_mycobank_db()}
}
\value{
Character vector. Taxonomic ranks separated by commas in each element.
}
\description{
Parses through the Mycobank database to return taxonomic classifications for a list of genera and/or genus sp. names
Names should be in a character vector and in the format "genus species" or just "genus" (variants and strains accepted)
If a given taxonomic name query consists of only one 'word' then taxonomic level of 'genus' is assumed
}
\details{
This will search through the database using pattern matching to pull the various basionyms, and synonyms for a given taxon name, if present. Returns NA for that taxon otherwise. The 'tidy_results' element of the returned list contains a 'long-form' tibble with all supported synonym types. Taxa with "sp." in their name will return genus-level synonyms only.
}
\examples{
db <- get_mycobank_db(overwrite = FALSE)
taxa <- c("Abaphospora borealis","Conisphaeria borealis var. minor","Sphaeria borealis","Nonsense name","Abaphospora sp.")
get_mycobank_taxonomy(taxa,db)

}
